Added new class ArtifactItemFilters
